The "Importance" report was moved to separate file.
#63
