+++ www/ExpensesAzure/Expenses.csproj
@@ -245,6 +245,7 @@
     <Compile Include="Controllers\MobileControllerDocuments.cs" />
     <Compile Include="Controllers\MobileControllerIncome.cs" />
     <Compile Include="Controllers\MobileControllerReports.cs" />
+    <Compile Include="Controllers\MobileControllerReportsImportance.cs" />
     <Compile Include="Controllers\MobileControllerReportsTop10.cs" />
     <Compile Include="Controllers\PersonalizedController.cs" />
     <Compile Include="Global.asax.cs">
